refactor(common): consolidate taxonomy term schema into a single source
Move the drizzle-zod taxonomy term encoder into @op/common so the decision
schemas can reuse it via .pick() instead of duplicating the shape. Tightens
termUri typing while collapsing two parallel definitions into one.
