+++ packages/common/src/services/decision/schemas/index.ts
@@ -4,4 +4,5 @@ export * from './instanceData';
 export * from './reviews';
 export * from './proposal';
 export * from './adminDecisionInstance';
+export * from './taxonomyTerm';
 export * from './transitionData';
