+++ packages/common/src/services/decision/schemas/taxonomyTerm.ts
@@ -1,0 +1,11 @@
+import { z } from 'zod';
+
+import { taxonomyTermsEncoder } from '../../terms/schemas';
+
+export const taxonomyTermSchema = taxonomyTermsEncoder.pick({
+  id: true,
+  label: true,
+  termUri: true,
+});
+
+export type TaxonomyTerm = z.infer<typeof taxonomyTermSchema>;
